requirements.txt
python-amazon-sp-api[orders]
fastapi
uvicorn[standard]
python-dotenv